Re-include the `target/doc` directory into the Git repository.
In addition to providing a bit more organization and clarity around
which filesystem paths are excluded from version control, this change
also ensures that the single `doc/` subdirectory within the `target/`
directory is included in commits. This is done to test whether we can
generate API documentation automatically while ensuring it remains local
to the code that it is documenting.