Wrap the Vercel Project's `installCommand` into a Makefile target.
Vercel Projects require that installation commands for their deployments
are no longer than 256 characters, yet we need a bit more in order to
actually build the project and generate API documentation on their build
image, which is an Amazon Linux 2 base. This commit wraps the install
command I've developed into a Makefile target so that we can simply use
a command such as `make vercel-api-docs` in the Vercel Project settings.