Fix "home mismatch" error during `rustup` installation forcibly.
This deals with a quirk in the Vercel.com build image where the `HOME`
environment variable is set differently from the result of calling the
functions `getpwuid(geteuid())` like how the `rustup` installer does.

Here we simply create a new `.PHONY` target to install `rustup` for any
Vercel image differently, but retains generic installation usage of it
so other recipes can reuse the actual Rust installation as a prereq.

See also:

* https://github.com/rust-lang/rustup/issues/1884