Working proof of concept for generated API docs automatically publishable to Vercel Project. (#373)
* Re-include the `target/doc` directory into the Git repository.

In addition to providing a bit more organization and clarity around
which filesystem paths are excluded from version control, this change
also ensures that the single `doc/` subdirectory within the `target/`
directory is included in commits. This is done to test whether we can
generate API documentation automatically while ensuring it remains local
to the code that it is documenting.

* Ensure the `target/doc` directory exists in the repository.

This is used simply to ensure that we can track the "empty" `target/doc`
directory as part of the repository contents. We want this included
because the directory is the root directory of automatically generated
API documentation for the project, used by downstream consumers.

* Wrap the Vercel Project's `installCommand` into a Makefile target.

Vercel Projects require that installation commands for their deployments
are no longer than 256 characters, yet we need a bit more in order to
actually build the project and generate API documentation on their build
image, which is an Amazon Linux 2 base. This commit wraps the install
command I've developed into a Makefile target so that we can simply use
a command such as `make vercel-api-docs` in the Vercel Project settings.

* Vercel doesn't provide a `sudo` command. Sad pandas. Let's just try.

* Install `clang-libs` dependency for Amazon Linux 2 in Vercel, too.

* Try with `clang-devel` as well to solve `stdarg.h` file not found.

* Use `rustup` instead of the Amazon Linux extras to install Rust.

* On Vercel, `$HOME` is weird, try not requiring it.

* Fix "home mismatch" error during `rustup` installation forcibly.

This deals with a quirk in the Vercel.com build image where the `HOME`
environment variable is set differently from the result of calling the
functions `getpwuid(geteuid())` like how the `rustup` installer does.

Here we simply create a new `.PHONY` target to install `rustup` for any
Vercel image differently, but retains generic installation usage of it
so other recipes can reuse the actual Rust installation as a prereq.

See also:

* https://github.com/rust-lang/rustup/issues/1884

* Replace `.gitkeep` with a nicer "API docs homepage."

* The only file we need in the `target/doc` directory is the `index.html`.

* Create a Vercel-specific build profile that optimizes for small size.

* Let `rustup` manage all tools in Vercel environment. Silences warnings.

* Build standard library with the Rust nightly we're using.

* Format `Cargo.toml` file to pass lint checks with `taplo fmt`.

* Move the complexity of the build command to the Makefile.

* Switch out PAT in favor of SSH deploy key.