Developer experience improvements: SSH auth from workstations, entirely local "devnet" functionality with Compose (#434)
* Reorder, sort excluded items listed in `.dockerignore`.

* Provide a "local devnet" setup with Docker Compose.

* Allow either SSH or GitHub PAT as authentication method while building.

This commit introduces the possibility of using a developer's local SSH
agent as part of the build process. It therefore also removes the hard
requirement of a "`credentials`" secret being available to the builder.

The change merely provides flexibility. Developers usually prefer to use
their own SSH credentials to authenticate to the remote repository, but
many CI/CD systems have more friction when doing this. This enhancement
simply lets us continue to use a single, canonical Dockerfile but still
create a situation where as many consumers of it are as happy as I can
reasonably make them, for the time being. <3

We also introduce a hardcoded GitHub.com SSH key (the ED25519 one), in
order to avoid any Trust On First Use (TOFU) vulnerability associated
with making connections to SSH servers impersonating GitHub.com's own.

* Use SSH agent forwarding when running `docker compose` as well.

* We don't use IPv6 (yet?) in the Docker Network, don't even ask for it.

* Tags in the `test/` tagspace should also be auto-marked as prerelease.