Add the benchmark template from the Polkadot v0.9.43 release
I've been using this in a couple of our PRs to generate weights, so figured it
would be good to actually commit to version control.

Once we update to Polkadot 1.0 and whatnot we'll probably want to follow their
templates as well so it gives us a place to store those.