Work clean. <3 <3 <3
This introduces a bit of development tooling: automatic Rust source code
checking with `rustfmt(1)`, just so that we can automate notions of code
style that conform to Rust's own style guides and not think much of it.

The same is done for `.toml` files with `taplo`'s `format` subcommand.

The script here is intended to be run as a Git pre-commit hook, and it's
placed into the `scripts/git/hooks` directory, which is defined by the
[`entropy-workstation-setup` utility] to be the canonical place for
shared Git hooks. You can configure repositories on a one-by-one basis:

```shell
cd $ENTROPY_WORKSPACE_DIR/entropy-core
git config --local core.hooksPath scripts/git/hooks
```

Or, preferably, you can simply re-run the `entropy-workstation-setup`
Playbooks. It knows to check for the existence of this directory in the
repositories it clones and sets repository configuration accordingly. It
also checks for the presence of both `rustfmt` and `taplo` at specific
versions and ensures they are available in Bash and Zsh search paths.

This script was written with the intent of being a developer aid, not an
obstacle or nuisance, and so it provides rather explanatory messaging in
the event one of its checks detects a formatting/linting concern.

[`entropy-workstation-setup` utility]: https://github.com/entropyxyz/entropy-workstation-setup