Provide a local development ("local devnet") configuration as a start. (#446)
* Provide a local development ("local devnet") configuration as a start.

This (finally?) provides a "local devnet" configuration intended to
replace the need for developers who are working with multiple components
of an Entropy network (in development/testing mode) to handle those
components independently. I.e., instead of invoking the `entropy` and
`server` binaries individually, the entire build and start up routine is
now encapsulated inside a `docker compose build && docker compose up`
invocation. If this works well, this will also become the basis on which
tests in CI can spin up an Entropy network to test against, as well.

There are a number of things to note about this change, most of which
are thoroughly commented inline, but some that deserve special call out:

1. Static comilation with GNU libc is not actually fully static. :(
    We still need to include the `libnss_files` and `libnss_dns` shared
    objects from the container image's `build` stage in order for the
    `entropy` binary to successfully make DNS queries. This wasn't
    necessary before, because those code paths were never called by an
    Entropy network only using IP addresses (or `127.0.0.1`, locally).
    However, the `local-devnet` chain introduced here cannot know IP
    addresses ahead of time, so we need name resolution and thus hit
    this very, very annoying bug.
1. The new `--chain local-devnet` option introduced here is specifically
    for use in combination with the Docker Compose `docker-compose.yaml`
    configuration provided here, as well. It'll fail with another setup.

As an aside, this is my first Rust contribution. Be gentle. <3

* Run `rustfmt` on the code in `node/cli/src/`, no functional changes.

* Only print the error hints if the commit is not good to go.