Change container distro to Debian and Ubuntu to avoid runtime errors. (#481)
This commit switches the container's distribution to Debian for building
and to Ubuntu for final distribution, because this more closely matches
upstream Substrate's build pipeline and seems to more reliably avoid
runtime issues caused by missing shared libraries and the like. We also
no longer compile statically, for now, as it's not needed in this setup,
and we try to maintain a smaller container image size by installing only
required packages and components of the Rust toolchain.

In my testing, this creates image sizes of just about ~200MB, which is
not terrible but not ideal (previous image sizes were down to ~60MB).
Nevertheless, as this also unblocks `entropy-js`, I think it's worth it.