Export plain binaries, always `strip(1)`, use `cargo build`.
This commit handles a recent modification in the `server` binary that
converted it into a Rust library. We can't use `rustc` to build this any
longer because the library needs to be linked against its binary in a
separate step. To do this, we have to use `cargo build` instead.

Additionally, we want to export the plain binaries (not as part of a
container image), so that we can post them as release artifacts when we
create a formal GitHub Release object for a given tagged version.