Add `.await` to all test logger calls
This isn't the nicest thing to do since we're now kind of assuming that all of our
test run in an async context.

However, I decided to make this choice since it would ensure that the Loki setup
logic was better encapsulated and the Loki background task would always be spawned.

At the moment all our of tests using the logger do run in an async context anyways,
so the tradeoff seemed okay.