Refactor Rust-based chain specs (#592)
* Move some chain spec accounts to constants

* Create new `chain_spec` module

* Fill out `dev` module

* Add WIP `load_spec` changes

* Fill out `testnet` module

* Add `testnet_local` configuration

* Remove `admin` module

We don't need this anymore. Functionality has been moved into the `chain_spec` module.

* Remove uneccessary chain configs

Let's keep this as simple as we can. If we need to add more chain configs later let's do that, but
only if they're acutally being used.

* Clean up `--chain` options

This make `--dev` the default option if no chain is passed in. In the future we'll may want to
change this to be our mainnet.

* TaploFmt

* Move testing configuration to own module

Ideally I'd want this to be the same as the `dev` config, but let's leave it this way for now.

* Rename `test` chainspec to `integration_tests`

* Move X25519 public keys into consts and own module

* Move TSS AccountID into own module

* Use new consts in `integration_tests` module

* Clean up some unused re-exports

* Use full word for "properties"

* Use `get_account_id_from_seed` from import instead of full path

* Add `Charlie` X25519 public key

* Update comments around what each network is made up of

* It doesn't need to be a validator for tests to pass

* Clean up unused re-export

* Update some comments

* Replace more hex accounts with named accounts

* Pull out `Eve` X25519 public key into const

* Add `devnet-local` configuration back

This is meant to be compatible with the hostname scheme expected by the local `docker-compose`
setup.

* Change `testnet-local` address to use Docker addresses

* Make note about breaking changes in `CHANGELOG`