Reduce size of council and technical committee
Since we added more accounts to the `endowed_accounts` list, and we use this list to come up with
our elections and technical committee members, we ended up hitting the limit on the member count.
Instead of upping the member count in the runtime config we reduced the number of members chosen at
genesis.