Use published version of Program crates (#781)
Starting with `v0.9.0` the `entropy-programs-core and `entropy-programs-runtime` crates are
available on crates.io. This PR uses those instead of using Git references.

I belive this is the last crate we need to update before the repo is in a theorically publishable
state.

Note that we still have the `template-barebones` crate as a Git dev-dependency, but it doesn't
impact publishing plans.