Fix release workflow by including only intended artifacts.
This commit fixes an issue introduced in #895 where Dependabot updated
the `docker/build-push-action` Action to major version 6 from 5. This
change [introduced the inclusion of Docker Build
Summaries](https://github.com/docker/build-push-action/releases/tag/v6.0.0),
which are build artifacts that end in `.dockerbuild` and are uploaded by
default. This change means we need to be more specific regarding the
artifacts we want to download and prepare for inclusion in a GitHub
Release itself. The build summaries are useful to have (so we don't turn
them off), but we don't want them included in the GitHub Release itself.