remove timeout for cargo test step
This was originally added as an analog to the circle-ci's "no_output_timeout",
but they are not actually analogs.  circle-ci's step measures how long
to wait without any output, while gh's timeout is how long to run the
step in general before timing out, regardless of output.  This could
cause the job to timeout when it was still successfully running, and
still within github's usage limits.