Migrate circle-ci workflow to github actions (#991)
* Add analogous gh jobs for .circleci/then.yml jobs

* Update .github/actions/install-dependencies/action.yaml

Co-authored-by: Hernando Castano <HCastano@users.noreply.github.com>

* Update .github/actions/install-dependencies/action.yaml

Co-authored-by: Hernando Castano <HCastano@users.noreply.github.com>

* Update .github/workflows/build-and-run-node-test.yaml

Co-authored-by: Hernando Castano <HCastano@users.noreply.github.com>

* Update .github/workflows/build-and-run-node-test.yaml

Co-authored-by: Hernando Castano <HCastano@users.noreply.github.com>

* remove timeout for cargo test step

This was originally added as an analog to the circle-ci's "no_output_timeout",
but they are not actually analogs.  circle-ci's step measures how long
to wait without any output, while gh's timeout is how long to run the
step in general before timing out, regardless of output.  This could
cause the job to timeout when it was still successfully running, and
still within github's usage limits.

* Remove circle-ci jobs

All relevant workflows have been ported to github actions.

* Run build-test on xl runner + small readme change

* Update .github/workflows/build-and-run-node-test.yaml

Co-authored-by: Hernando Castano <HCastano@users.noreply.github.com>

---------

Co-authored-by: Hernando Castano <HCastano@users.noreply.github.com>
Co-authored-by: Hernando Castano <hernando@hcastano.com>