Use loose coupling between attestation and staking extension pallets
With the tight coupling we're not able to build out a mock runtime for the staking extension pallet
since the compiler isn't able to verify that we have an implementation of the staking extension
pallet for our mock attestation pallet (yeah, bit of a roundabout thing...).

The loose coupling gives us a bit of flexibility here, and also allows us to scope the iteraction
between the pallets in a more fine grained way.