Input expected keys into quote verification method directly
We need to match what the TSS is committing to (e.g what we write to storage) on the Staking pallet
end, but we don't always have that information available on the Attestation pallet side. By having
the expected data come in alongside the quote we can more easily check that it does match.