With custom chain inputs, we no longer use a pre-generated chainspec.
This commit removes the pre-generated partial chainspec file from being
generated by the `entropy-core` CI systems on builds. This file was
necessary when we were patching the chainspec file as part of a standard
network deployment but is no longer needed with the new ability to
simply provide custom inputs that get used in the chainspec file,
effectively moving the patching of this file into core.

See also: #1424, and #1447.